tests/test_with_python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test_with_python"
version = "0.1.0"
description = "Test Python client for RMCP"
requires-python = ">=3.12"
dependencies = [
  "fastmcp==2.10.1",
]

[tool.setuptools]
py-modules = ["client", "server"]
